debug

+++ modules/controllers/ProductController.php
@@ -154,16 +154,16 @@ class ProductController extends Controller
 //        die;
         $returnValue = '';
 //        if(file_exists($image))
-        {
-            $output = [];
-            $path = \Yii::$app->params['dir_path'];
-//            if(file_exists($path."vtour"))
-//            {
-                (exec("rm -rf /tmp/vr/vtour"));
-//            }
-            (exec($command,$output,$returnValue));
-        }
 
+        $output = [];
+        $path = \Yii::$app->params['dir_path'];
+        if(file_exists($path."vtour"))
+        {
+            (exec("rm -rf /tmp/vr/vtour"));
+        }
+        (exec($command,$output,$returnValue));
+
+var_dump($returnValue);die;
         if($returnValue === 0)
         {
             //拷贝panos dir
@@ -174,8 +174,8 @@ class ProductController extends Controller
 //            @copy($soucre,$dest);
             $execute_cp = "cp -rf ".$soucre." ".$dest;
             exec($execute_cp);
-            $execute = "chmod  -R 777 ".\Yii::$app->basePath."/web/" ."vtour";
-            exec($execute);
+//            $execute = "chmod  -R 777 ".\Yii::$app->basePath."/web/" ."vtour";
+//            exec($execute);
 
             //todb and genxml
             $this->actionConv($proId);
